main/RAG/requirements.txt
qdrant-client
langchain
langchain-community
langchain-core
langchain-qdrant
langchain-openai
langchain-groq
langchain-huggingface
python-dotenv
fastembed
sentence-transformers
unstructured
unstructured[pdf]
pypdf
boto3
tiktoken
numpy
groq